src/cloud-services/shared-services/budgeting-rate-limiting-cost-observability/requirements.txt
# Dependencies for Budgeting, Rate-Limiting & Cost Observability Module (M35)
# Per PRD v3.0.0

# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# Redis for caching and rate limiting
redis>=5.0.0
hiredis>=2.2.0

# Cryptography for receipt signing
cryptography>=41.0.0

# HTTP client (for testing and external APIs)
httpx>=0.25.0

# JSON handling
orjson>=3.9.0

# Date/time handling
python-dateutil>=2.8.0

# JWT handling (for IAM integration)
PyJWT>=2.8.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Performance testing
locust>=2.17.0

# Monitoring and metrics
prometheus-client>=0.19.0

